+++ HidingBar/HidingBar.lua
@@ -54,7 +54,7 @@ if MSQ then
 			hidingBar:MSQ_CoordUpdate(btn)
 		end
 		hidingBar:enter()
-		hidingBar:leave(1.5)
+		hidingBar:leave(math.max(1.5, hidingBar.config.hideDelay))
 	end)
 
 
@@ -66,7 +66,7 @@ if MSQ then
 			hidingBar:MSQ_CoordUpdate(btn)
 		end
 		hidingBar:enter()
-		hidingBar:leave(1.5)
+		hidingBar:leave(math.max(1.5, hidingBar.config.hideDelay))
 	end)
 
 
@@ -148,11 +148,14 @@ if MSQ then
 					data._Icon:SetAtlas(atlas)
 				end
 				data._Normal.SetTexture = function(_, texture)
-					data._Icon:SetTexture(texture)
+					if texture then
+						data._Icon:SetTexture(texture)
+					end
 				end
 				data._Normal.SetTexCoord = function(_, ...)
 					data._Icon:SetTexCoord(...)
 				end
+				data._Normal = nil
 			end
 			if data._Pushed then
 				data._Pushed:SetAlpha(0)
@@ -550,7 +553,7 @@ end
 function hidingBar:grabMinimapAddonsButtons(parentFrame)
 	for _, child in ipairs({parentFrame:GetChildren()}) do
 		local width, height = child:GetSize()
-		if max(width, height) > 16 and math.abs(width - height) < 5 then
+		if math.max(width, height) > 16 and math.abs(width - height) < 5 then
 			self:addMButton(child)
 		end
 	end
@@ -1347,9 +1350,12 @@ function hidingBar:refreshShown()
 end
 
 
-function hidingBar.drag:hoverHandler()
+function hidingBar.drag:hoverWithClick()
 	if self:IsShown() and hidingBar.config.fade then
 		UIFrameFadeOut(self, hidingBar.config.showDelay, self:GetAlpha(), 1)
+	end
+	if hidingBar:IsShown() then
+		hidingBar:enter()
 	end
 end
 
@@ -1367,7 +1373,9 @@ function hidingBar.drag:showOnHoverWithDelay()
 	if hidingBar:IsShown() or hidingBar.config.showDelay == 0 then
 		hidingBar:enter()
 	else
-		self:hoverHandler()
+		if self:IsShown() and hidingBar.config.fade then
+			UIFrameFadeOut(self, hidingBar.config.showDelay, self:GetAlpha(), 1)
+		end
 		self.fTimer.timer = hidingBar.config.showDelay
 		self.fTimer:SetScript("OnUpdate", self.showBarDelay)
 	end
@@ -1384,7 +1392,7 @@ function hidingBar.drag:setShowHandler(showHandler)
 		self:SetScript("OnEnter", self.showOnHoverWithDelay)
 		self:SetScript("OnClick", enter)
 	elseif hidingBar.config.showHandler == 1 then
-		self:SetScript("OnEnter", self.hoverHandler)
+		self:SetScript("OnEnter", self.hoverWithClick)
 		self:SetScript("OnClick", enter)
 	else
 		self:SetScript("OnEnter", self.showOnHoverWithDelay)